work on build and structure
	pick up build/version number from `package.json` version
	move actual source code to `src` directory
	move all images to `images` directory, update references
	change stray `.js` file to `.ts`
	add prettier config
	update README
	style: tighten up call-to-action ("Enter price in USD") heading somewhat
