--- index.html
@@ -1,50 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <!--
-
-    Web app developed and maintained by Raven Rebels 
-    https://www.ravenrebels.com
-    https://twitter.com/RavenRebels
-
-    Info about background image.
-    Creator: Circe Denyer, https://linktr.ee/circod
-    URL: https://www.publicdomainpictures.net/en/view-image.php?image=261251&picture=raven-silhouette-and-storm-clouds
-
-
-  -->
-  <head>
-    <meta name="apple-mobile-web-app-capable" content="yes" />
-    <meta charset="UTF-8" />
-    <title>RVN price calculator</title>
-
-    <link rel="apple-touch-icon" sizes="180x180" href="apple-touch-icon.png" />
-    <link rel="icon" type="image/png" sizes="32x32" href="favicon-32x32.png" />
-    <link rel="icon" type="image/png" sizes="16x16" href="favicon-16x16.png" />
-    <link rel="manifest" href="manifest.json" />
-
-    <!-- prevent caching start -->
-    <meta
-      http-equiv="Cache-Control"
-      content="no-cache, no-store, must-revalidate"
-    />
-    <meta http-equiv="Pragma" content="no-cache" />
-    <meta http-equiv="Expires" content="0" />
-    <!-- prevent caching end -->
-
-    <link
-      href="https://fonts.googleapis.com/css?family=Quicksand:300,500"
-      rel="stylesheet"
-    />
-    <link
-      href="https://fonts.googleapis.com/css2?family=B612+Mono&display=swap"
-      rel="stylesheet"
-    />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-  </head>
-  <body>
-    <div id="app"></div>
-    <script type="module" src="./index.js?asdf"></script>
-    <link rel="stylesheet" href="./style.css?asdf" />
-    <link rel="apple-touch-icon" href="./raven.png" />
-  </body>
-</html>
