+++ snowpack.config.js
@@ -1,0 +1,10 @@
+// Snowpack Configuration File
+// See all supported options: https://www.snowpack.dev/reference/configuration
+
+/** @type {import("snowpack").SnowpackUserConfig } */
+export default {
+    mount: {
+        public: "/",
+        src: "/",
+    },
+};
