+++ src/images/android-chrome-192x192.png
