+++ src/images/favicon-16x16.png
