+++ src/index.ts
@@ -1,0 +1,10 @@
+const url = "https://api1.binance.com/api/v3/ticker/price?symbol=RVNUSDT";
+
+async function fetchPrice(): Promise<number> {
+    const response = await fetch(url);
+    const priceObject = await response.json();
+    return priceObject.price;
+}
+
+//Update each minute
+setInterval(fetchPrice, 60 * 1000);
