+++ src/manifest.json
@@ -1,0 +1,25 @@
+{
+    "name": "USD to RVN calculator",
+    "short_name": "USD to RVN calculator",
+    "start_url": "index.html",
+    "display": "standalone",
+    "background_color": "#fff",
+    "description": "RVN price calculator",
+    "icons": [
+        {
+            "src": "/images/raven.png",
+            "sizes": "48x48",
+            "type": "image/png"
+        },
+        {
+            "src": "/images/android-chrome-192x192.png",
+            "sizes": "192x192",
+            "type": "image/png"
+        },
+        {
+            "src": "/images/android-chrome-512x512.png",
+            "sizes": "512x512",
+            "type": "image/png"
+        }
+    ]
+}
