+++ src/style.css
@@ -1,0 +1,88 @@
+body {
+    background-repeat: no-repeat;
+    background-size: auto;
+    background-color: #000;
+    background-image: linear-gradient(147deg, #923cb5 0%, #000000 74%);
+    color: #fff;
+    display: flex;
+    font-family: "Quicksand", sans-serif;
+    font-size: 1.5rem;
+    font-weight: 300;
+    justify-content: center;
+    margin: 0;
+    max-width: calc(100vw);
+    min-height: calc(100vh);
+    padding: 0;
+    padding: 0;
+}
+h1,
+h2,
+h3,
+h4,
+h5,
+h6 {
+    font-weight: 500;
+}
+.ravencoin-price-calculator__main-container {
+    margin-top: 20px;
+    padding-top: 10px !important;
+}
+@media only screen and (min-width: 600px) {
+    .ravencoin-price-calculator__main-container {
+        margin-top: 30px;
+        padding-top: 20px;
+    }
+}
+
+#app {
+    max-width: calc(90vw);
+}
+
+.ravencoin-price-calculator__table td {
+    width: 100px;
+    font-family: "B612 Mono", monospace;
+}
+.ravencoin-price-calculator__table th {
+    text-align: left;
+}
+.ravencoin-price-calculator__table td:nth-child(2) {
+    text-align: right;
+}
+.ravencoin-price-calculator__glass-container {
+    backdrop-filter: blur(1px);
+    color: white;
+    height: auto;
+    min-width: 100px;
+    padding: 2rem;
+    background: rgba(255, 255, 255, 0.1);
+    border-radius: 10px;
+    border: 1px solid rgba(255, 255, 255, 0.18);
+}
+
+.ravencoin-price-calculator__subtitle {
+    display: flex;
+    margin-bottom: 20px;
+    justify-content: space-between;
+}
+.ravencoin-price-calculator__cta {
+    margin: 0 0 1rem;
+}
+.ravencoin-price-calculator__input {
+    border: 0;
+    border-radius: 10px;
+    padding: 0.3rem;
+    font-size: 18px;
+    display: block;
+    margin-bottom: 10px;
+    width: 200px;
+}
+.ravencoin-price-calculator__footer {
+    font-size: 16px;
+    color: white;
+    margin-top: 40px;
+    padding: 20px;
+    padding-left: 0;
+}
+.ravencoin-price-calculator__footer-link {
+    color: white;
+}
